Removed models from study to dashboard (temp) registered the models
