+++ study/models.py
@@ -62,18 +62,3 @@ class Progress(models.Model):
 def assign_test_id():
 		return Test.objects.all().count()+1	
 '''
-class TestData(models.Model):
-	test_id=models.IntegerField(default=1)
-	user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
-	word_id= models.ForeignKey(WordList,on_delete=models.CASCADE)
-	right= models.BooleanField(default=False)
-	
-	
-
-class Test(models.Model):
-	test_data=models.ForeignKey(TestData, on_delete=models.CASCADE)
-	test_date= models.DateTimeField(default=timezone.now)
-
-	def getscore(self):
-		return self.test_data.objects.filter(right=True).count()
-
